address review comments, make initial dealy of scheduling job to start to 10min in purpose of waiting cluster members to be all up
